Retry deletion of failed instances
If an instance fails to delete for some reason, we leave it in error state.
This allows people to see that there is something wrong.
However, that instance is never removed until the server is restarted. This way it take one slot in TeamCity and one less instance can be created.

Create a background job that will check for failed instances every 5 min and will retry to delete them.