Use built-in scheduled executor
We had our only scheduled executor, but it had limitations.
As we have 2 threads by default and those threads are used for creation and deletion of VMs, it means we can do 2 CRUD operations at the same time.
Use the built-in executor instead. We can use as many threads as there are available.